Rename collection_list to accounts_list
