+++ app/javascript/mastodon/features/collections/detail/accounts_list.tsx
