+++ app/javascript/mastodon/features/collections/detail/index.tsx
@@ -24,7 +24,7 @@ import { me } from 'mastodon/initial_state';
 import { fetchCollection } from 'mastodon/reducers/slices/collections';
 import { useAppDispatch, useAppSelector } from 'mastodon/store';
 
-import { CollectionAccountsList } from './collection_list';
+import { CollectionAccountsList } from './accounts_list';
 import { CollectionMetaData } from './collection_list_item';
 import { CollectionMenu } from './collection_menu';
 import classes from './styles.module.scss';
